+++ Program.cs
@@ -45,6 +45,7 @@ builder.Services.AddScoped<IReservePotActionService, ReservePotActionService>();
 builder.Services.AddScoped<IReserveInterestForecastService, ReserveInterestForecastService>();
 builder.Services.AddScoped<IFinancialForecastService, FinancialForecastService>();
 builder.Services.AddScoped<IWhatIfForecastService, WhatIfForecastService>();
+builder.Services.AddScoped<IForecastScenarioService, ForecastScenarioService>();
 builder.Services.AddScoped<FinanceRepository>();
 builder.Services.AddScoped<FinanceCalculator>();
 builder.Services.AddHttpClient<MarketPriceService>();